Move link plugin functionality to lexical directory
