+++ app/gui2/src/components/FloatingSelectionMenu.vue
