+++ app/gui2/src/components/lexical/LinkPlugin/__tests__/LinkPlugin.test.ts
