+++ app/gui2/src/components/lexical/LinkPlugin/autoMatcher.ts
