+++ app/gui2/src/components/lexical/LinkToolbar.vue
